Flytt autentiseringsfunksjoner inn i klientene
