+++ src/main/kotlin/rekrutteringsbistand/stilling/indekser/autentisering/authenticationUtils.kt
@@ -3,26 +3,7 @@ package rekrutteringsbistand.stilling.indekser.autentisering
 import com.github.kittinunf.fuel.core.FuelManager
 import com.github.kittinunf.fuel.core.extensions.authentication
 
-fun authenticateWithAzureAdToken(httpClient: FuelManager, accessTokenClient: AccessTokenClient): FuelManager {
-    val rekrutteringsbistandApiClientId = "fe698176-ac44-4260-b8d0-dbf45dd956cf"
-
-    addBearerToken(httpClient) {
-        accessTokenClient.getAccessToken(scope = "api://$rekrutteringsbistandApiClientId/.default")
-    }
-
-    return httpClient
-}
-
-fun authenticateWithElasticSearchCredentials(httpClient: FuelManager): FuelManager {
-    addBasicAuthentication(httpClient,
-            username = "todo",
-            password = "todo"
-    )
-
-    return httpClient
-}
-
-private fun addBearerToken(httpClient: FuelManager, getToken: () -> AccessToken) {
+fun addBearerToken(httpClient: FuelManager, getToken: () -> AccessToken) {
     httpClient.addRequestInterceptor {
         { request ->
             val token = getToken()
@@ -32,7 +13,7 @@ private fun addBearerToken(httpClient: FuelManager, getToken: () -> AccessToken)
     }
 }
 
-private fun addBasicAuthentication(httpClient: FuelManager, username: String, password: String) {
+fun addBasicAuthentication(httpClient: FuelManager, username: String, password: String) {
     httpClient.addRequestInterceptor {
         { request ->
             request.authentication().basic(username, password)
