+++ src/main/kotlin/rekrutteringsbistand/stilling/indekser/elasticsearch/ElasticSearchClient.kt
@@ -1,6 +1,7 @@
 package rekrutteringsbistand.stilling.indekser.elasticsearch
 
 import com.github.kittinunf.fuel.core.FuelManager
+import rekrutteringsbistand.stilling.indekser.autentisering.addBasicAuthentication
 
 class ElasticSearchClient(private val httpClient: FuelManager) {
     fun indekserStilling(stilling: Stilling) {
@@ -11,10 +12,18 @@ class ElasticSearchClient(private val httpClient: FuelManager) {
         const val esUrl: String = "todo"
         const val esIndex: String = "todo"
         const val esType: String = "todo"
+
+        fun authenticateWithElasticSearchCredentials(httpClient: FuelManager): FuelManager {
+            addBasicAuthentication(httpClient,
+                    username = "todo",
+                    password = "todo"
+            )
+
+            return httpClient
+        }
     }
 }
 
-// TODO: Ekspander type
 data class Stilling (
     val uuid: String,
     val title: String,
