+++ src/main/kotlin/rekrutteringsbistand/stilling/indekser/stillingsinfo/StillingsinfoClient.kt
@@ -4,6 +4,8 @@ import com.github.kittinunf.fuel.core.FuelManager
 import com.github.kittinunf.fuel.jackson.responseObject
 import com.github.kittinunf.result.Result
 import org.eclipse.jetty.http.HttpStatus
+import rekrutteringsbistand.stilling.indekser.autentisering.AccessTokenClient
+import rekrutteringsbistand.stilling.indekser.autentisering.addBearerToken
 import rekrutteringsbistand.stilling.indekser.environment
 import java.lang.RuntimeException
 
@@ -31,6 +33,16 @@ class StillingsinfoClient(private val httpClient: FuelManager) {
 
     companion object {
         val stillingsinfoUrl: String = "${environment().get("REKRUTTERINGSBISTAND_API")}/indekser/stillingsinfo"
+
+        fun authenticateWithAzureAdToken(httpClient: FuelManager, accessTokenClient: AccessTokenClient): FuelManager {
+            val rekrutteringsbistandApiClientId = "fe698176-ac44-4260-b8d0-dbf45dd956cf"
+
+            addBearerToken(httpClient) {
+                accessTokenClient.getAccessToken(scope = "api://$rekrutteringsbistandApiClientId/.default")
+            }
+
+            return httpClient
+        }
     }
 }
